[enh #113] 作業必要ファイル追加
