+++ ProjectAzur/ProjectAzur/MainGame/GameScene/Player/PlayerAction/paPlayerFishAction.cpp
@@ -1,0 +1,36 @@
+﻿//==================================================================================================================================//
+//!< @file		paPlayerFishAction.cpp	
+//!< @brief		pa::PlayerFishActionクラスの実装
+//!< @author	T.Haga
+//!< @data		作成日時：2018/01/26	更新履歴：
+//==================================================================================================================================//
+
+/* Includes --------------------------------------------------------------------------------------------------- */
+
+#include "paPlayerFishAction.h"
+
+namespace pa
+{
+
+/* Public Functions ------------------------------------------------------------------------------------------- */
+
+PlayerFishAction::PlayerFishAction()
+{}
+
+PlayerFishAction::~PlayerFishAction()
+{}
+
+void PlayerFishAction::Initialize()
+{}
+
+void PlayerFishAction::Update(PlayerParam& rParam)
+{}
+
+void PlayerFishAction::Finalize()
+{}
+
+}	// namespace pa
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
