+++ ProjectAzur/ProjectAzur/MainGame/GameScene/Player/PlayerAction/paPlayerOriginalAction.cpp
@@ -1,0 +1,37 @@
+﻿//==================================================================================================================================//
+//!< @file		paPlayerOriginalAction.cpp		
+//!< @brief		pa::PlayerOriginalActionクラスの実装
+//!< @author	T.Haga
+//!< @data		作成日時：2018/01/26	更新履歴：
+//==================================================================================================================================//
+
+/* Includes --------------------------------------------------------------------------------------------------- */
+
+#include "paPlayerOriginalAction.h"
+
+namespace pa
+{
+
+/* Public Functions ------------------------------------------------------------------------------------------- */
+
+PlayerOriginalAction::PlayerOriginalAction()
+{}
+
+PlayerOriginalAction::~PlayerOriginalAction()
+{}
+
+void PlayerOriginalAction::Initialize()
+{}
+
+void PlayerOriginalAction::Update(PlayerParam& rParam)
+{}
+
+void PlayerOriginalAction::Finalize()
+{}
+
+
+}	// namespace pa
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
